add StandaloneReverseRegistrar abi and refactor namespaces into namespaces.ts
