+++ packages/datasources/src/namespaces.ts
@@ -1,0 +1,56 @@
+import ensTestEnv from "./ens-test-env";
+import holesky from "./holesky";
+import { DatasourceNames, ENSNamespace, ENSNamespaceId } from "./lib/types";
+import mainnet from "./mainnet";
+import sepolia from "./sepolia";
+
+// internal map ENSNamespaceId -> ENSNamespace
+const ENSNamespacesById = {
+  mainnet,
+  sepolia,
+  holesky,
+  "ens-test-env": ensTestEnv,
+} as const satisfies Record<ENSNamespaceId, ENSNamespace>;
+
+/**
+ * Returns the ENSNamespace for a specified `namespaceId`.
+ *
+ * @param namespaceId - The ENSNamespace identifier (e.g. 'mainnet', 'sepolia', 'holesky', 'ens-test-env')
+ * @returns the ENSNamespace
+ */
+export const getENSNamespace = <N extends ENSNamespaceId>(
+  namespaceId: N,
+): (typeof ENSNamespacesById)[N] => ENSNamespacesById[namespaceId];
+
+/**
+ * Returns the `datasourceName` Datasource within the specified `namespaceId` namespace.
+ *
+ * NOTE: the typescript typechecker _will_ enforce validity. i.e. using an invalid `datasourceName`
+ * within the specified `namespaceId` will be a type error.
+ *
+ * @param namespaceId - The ENSNamespace identifier (e.g. 'mainnet', 'sepolia', 'holesky', 'ens-test-env')
+ * @param datasourceName - The name of the Datasource to retrieve
+ * @returns The Datasource object for the given name within the specified namespace
+ */
+export const getDatasource = <
+  N extends ENSNamespaceId,
+  D extends keyof ReturnType<typeof getENSNamespace<N>>,
+>(
+  namespaceId: N,
+  datasourceName: D,
+) => getENSNamespace(namespaceId)[datasourceName];
+
+/**
+ * Returns the chain for the ENS Root Datasource within the selected namespace.
+ *
+ * @returns the chain that hosts the ENS Root
+ */
+export const getENSRootChain = (namespaceId: ENSNamespaceId) =>
+  getDatasource(namespaceId, DatasourceNames.ENSRoot).chain;
+
+/**
+ * Returns the chain id for the ENS Root Datasource within the selected namespace.
+ *
+ * @returns the chain ID that hosts the ENS Root
+ */
+export const getENSRootChainId = (namespaceId: ENSNamespaceId) => getENSRootChain(namespaceId).id;
